+++ src/compas_timber/structural.py
@@ -1,3 +1,6 @@
+from __future__ import annotations
+
+import typing
 from typing import List
 
 from compas.data import Data
@@ -6,9 +9,12 @@ from compas.geometry import Point
 from compas.geometry import closest_point_on_segment
 from compas.geometry import distance_point_point
 from compas.itertools import pairwise
+from compas.tolerance import TOL
 
-from compas_timber.elements import Beam
-from compas_timber.model import TimberModel
+if typing.TYPE_CHECKING:
+    from compas_timber.connections import Joint
+    from compas_timber.elements import Beam
+    from compas_timber.model import TimberModel
 
 
 class StructuralSegment(Data):
@@ -26,10 +32,12 @@ class StructuralSegment(Data):
 
 
 class StructuralElementSolver:
-    def create_structural_segments(self, beam: Beam, model: TimberModel) -> List[StructuralSegment]:
-        """Create structural segments for a given beam in the timber model.
+    """Produces structural segments for beams and joints in a timber model."""
 
-        This are essentially segments of the beam's centerline split at the locations of the joints.
+    def add_structural_segments(self, beam: Beam, model: TimberModel) -> List[StructuralSegment]:
+        """Creates and adds structural segments for a given beam to the timber model.
+
+        These are essentially segments of the beam's centerline split at the locations of the joints.
 
         Parameters
         ----------
@@ -38,18 +46,24 @@ class StructuralElementSolver:
         model : :class:`compas_timber.model.TimberModel`
             The timber model containing the beams and joints.
 
-        Returns
-        -------
-        List[:class:`compas_timber.model.StructuralSegment`]
-            A list of structural segments created for the beam.
         """
-        joints = model.get_interactions_for_element(beam)
+        joints_for_beam = model.get_interactions_for_element(beam)
+        segments = self._create_segments(beam, joints_for_beam)
+        model.add_beam_structural_segments(beam, segments)
+        return segments
 
+    def _create_segments(self, beam: Beam, joints: List[Joint]) -> List[StructuralSegment]:
         # create segments between joints
         split_points_with_distances = []
         for joint in joints:
             point_on_segment = Point(*closest_point_on_segment(joint.location, beam.centerline))
             distance_from_start = distance_point_point(beam.centerline.start, point_on_segment)
+            distance_from_end = distance_point_point(beam.centerline.end, point_on_segment)
+
+            if TOL.is_zero(distance_from_start) or TOL.is_zero(distance_from_end):
+                # joints at start and end do not require splitting, as they are already segment boundaries
+                continue
+
             split_points_with_distances.append((distance_from_start, point_on_segment))
 
         # sort split point along the centeline